Set texture wrapping to repeat if not clamped
If the 'clamp' parameter of set_texture was False, no wrap was
explicitly set. And on clamp=2, the repeat wrap was set.
This makes no sense, as repeat should be the default.
So now it is changed to set repeat on clamp=False,
and to set clamped if clamp=True.